+++ src/main/java/kitchenpos/order/common/domain/OrderRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.order.common.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.order.eatinorder.domain.OrderTable;
 
 public interface OrderRepository {
     Order save(Order order);
