+++ src/main/java/kitchenpos/order/common/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.order.common.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
