+++ src/test/java/kitchenpos/order/common/application/OrderServiceTest.java
@@ -1,13 +1,16 @@
-package kitchenpos.application;
+package kitchenpos.order.common.application;
 
+import kitchenpos.order.devlieryorder.infra.FakeKitchenridersClient;
+import kitchenpos.order.common.infra.InMemoryOrderRepository;
+import kitchenpos.order.eatinorder.infra.InMemoryOrderTableRepository;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.OrderType;
+import kitchenpos.order.common.domain.Order;
+import kitchenpos.order.common.domain.OrderLineItem;
+import kitchenpos.order.common.domain.OrderRepository;
+import kitchenpos.order.common.domain.OrderStatus;
+import kitchenpos.order.eatinorder.domain.OrderTable;
+import kitchenpos.order.eatinorder.domain.OrderTableRepository;
+import kitchenpos.order.common.domain.OrderType;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
